+++ app/views/admin/discounts/edit.erb
