+++ app/views/admin/discounts/new.erb
@@ -2,7 +2,7 @@
     <aside class="sidebar">
         <ul class="menu-container">
             <li class="menu-option">
-                <a class="menu-link" href="/discounts">Back to discounts</a>
+                <a class="menu-link" href="/admin/discounts">Back to discounts</a>
             </li>
         </ul>
     </aside>
@@ -10,7 +10,7 @@
 
         <h2 id="item-type">New Discount</h2>
         <hr>
-        <form class="pure-form pure-form-aligned" action="/discounts" accept-charset="UTF-8" enctype="multipart/form-data" method="post">
+        <form class="pure-form pure-form-aligned" action="/admin/discounts" accept-charset="UTF-8" enctype="multipart/form-data" method="post">
             <%= csrf_tag %>
             <fieldset>
                 <div class="pure-control-group">
