+++ config/application.rb
@@ -92,7 +92,6 @@ class UsefulMusic::App
   controller '/items', ItemsController
   controller '/purchases', PurchasesController
   controller '/shopping_baskets', ShoppingBasketsController
-  controller '/discounts', DiscountsController
   controller '/orders', OrdersController
   controller '/about', AboutController
   controller '/admin', UsefulMusic::AdminController
